requirements.txt
bertopic
chardet
hdbscan
numpy
pandas>=2.0.3,<3.0
pandas-profiling>=3.6.6,<4.0
plotly
pyvis
requests
scikit-learn
sentence-transformers
streamlit>=1.31.1,<2.0
streamlit-highcharts>=0.1.5
streamlit_extras
streamlit_pandas_profiling
streamlit-condition-tree
openai
openpyxl
umap-learn
XlsxWriter
ydata-profiling
faiss-cpu